requirements.txt
Django
python-postmark>=0.4.3
pymongo
hashlib
datetime
